+++ QvaDev.Common/Integration/IFixConnector.cs
@@ -2,7 +2,7 @@
 {
 	public interface IFixConnector : IConnector
 	{
-		double SendMarketOrderRequest(string symbol, Sides side, decimal quantity, string comment = null);
+		decimal SendMarketOrderRequest(string symbol, Sides side, decimal quantity, string comment = null);
 		void OrderMultipleCloseBy(string symbol);
 		SymbolData GetSymbolInfo(string symbol);
 	}
