+++ QvaDev.Common/Integration/Tick.cs
@@ -4,8 +4,8 @@ namespace QvaDev.Common.Integration
 {
     public class Tick
     {
-        public double Ask { get; set; }
-        public double Bid { get; set; }
+        public decimal Ask { get; set; }
+        public decimal Bid { get; set; }
         public string Symbol { get; set; }
         public DateTime Time { get; set; }
     }
